Refactor switch(command) in loop() to a function pointer array
To streamline the main loop() and reduce the complexity introduced
by extensive switch-case structures, a function pointer array
approach was adopted. The incoming command byte from the PC serves
as an index into this array, directly mapping to the corresponding
function to be executed.

+++ CmdProcessor/src/adf4356.cpp
@@ -54,7 +54,7 @@ uint32_t ADF4356_LO::set_DLD() {
   return Curr.Reg[4];
 }
 
-uint32_t ADF4356_LO::ADF4356Execute(int commandIndex) {
+uint32_t ADF4356_LO::ADF4356Execute(byte commandIndex) {
   if (commandIndex >= 0 && commandIndex < ADF4356_LO::NUMBER_OF_FUNCTIONS) {
     return (this->*adfCmds[commandIndex])();
   }
